chore(suite-native): move timeswitch to separate component
