+++ suite-native/graph/src/hooks.ts
@@ -32,7 +32,6 @@ import {
     selectAccountGraphTimeframe,
     selectPortfolioGraphTimeframe,
     setAccountGraphTimeframe,
-    setPortfolioGraphTimeframe,
 } from './slice';
 import { type TimeframeHoursValue } from './types';
 import { omitErrorMessageSensitiveData } from './utils';
@@ -156,7 +155,6 @@ export const useGraphForSingleAccount = ({
 };
 
 export const useGraphForAllDeviceAccounts = ({ baseCurrencyCode }: CommonUseGraphParams) => {
-    const dispatch = useDispatch();
     // Use deep comparison because account items are rebuilt from account data.
     const accountItems = useSelectorDeepComparison(
         selectPortfolioGraphAccountItemsIfDiscoveryIsNotRunning,
@@ -168,15 +166,6 @@ export const useGraphForAllDeviceAccounts = ({ baseCurrencyCode }: CommonUseGrap
 
     const { startOfTimeFrameDate, endOfTimeFrameDate } =
         useGetTimeFrameForHistoryHours(portfolioGraphTimeframe);
-
-    const handleSelectPortfolioTimeframe = useCallback(
-        (timeframeHours: TimeframeHoursValue) => {
-            if (portfolioGraphTimeframe !== timeframeHours) {
-                dispatch(setPortfolioGraphTimeframe({ timeframeHours }));
-            }
-        },
-        [dispatch, portfolioGraphTimeframe],
-    );
 
     useWatchTimeframeChangeForAnalytics(portfolioGraphTimeframe);
 
@@ -195,7 +184,6 @@ export const useGraphForAllDeviceAccounts = ({ baseCurrencyCode }: CommonUseGrap
         ...graphForAccounts,
         isAnyMainnetAccountPresent: A.isNotEmpty(accountItems),
         timeframe: portfolioGraphTimeframe,
-        onSelectTimeFrame: handleSelectPortfolioTimeframe,
     };
 };
