+++ suite-native/graph/src/index.ts
@@ -5,3 +5,4 @@ export * from './hooks';
 export * from './slice';
 export * from './selectors';
 export * from './components/GraphBaseCurrencyBalance';
+export { type TimeframeHoursValue } from './types';
