+++ suite-native/module-home/src/screens/HomeScreen/components/PortfolioGraph.tsx
@@ -8,7 +8,6 @@ import { selectSelectedDeviceTotalFiatBalance } from '@suite-native/device';
 import { useIsDiscoveryDurationTooLong } from '@suite-native/discovery';
 import {
     Graph,
-    TimeSwitch,
     selectDeviceHistoryIgnoredNetworkSymbols,
     selectHasDeviceHistoryEnabledAccounts,
     useGraphAtoms,
@@ -17,6 +16,7 @@ import {
 import { Translation } from '@suite-native/intl';
 
 import { referencePointAtom, selectedPointAtom } from '../portfolioGraphAtoms';
+import { PortfolioGraphTimeSwitch } from './PortfolioGraphTimeSwitch';
 import { PortfolioHeader } from './PortfolioHeader';
 
 export type PortfolioGraphRef = {
@@ -50,17 +50,10 @@ export const PortfolioGraph = forwardRef<PortfolioGraphRef>((_props, ref) => {
     const loadingTakesLongerThanExpected = useIsDiscoveryDurationTooLong();
     const totalFiatBalance = useSelector(selectSelectedDeviceTotalFiatBalance);
 
-    const {
-        graphPoints,
-        error,
-        isLoading,
-        isAnyMainnetAccountPresent,
-        refetch,
-        onSelectTimeFrame,
-        timeframe,
-    } = useGraphForAllDeviceAccounts({
-        baseCurrencyCode,
-    });
+    const { graphPoints, error, isLoading, isAnyMainnetAccountPresent, refetch } =
+        useGraphForAllDeviceAccounts({
+            baseCurrencyCode,
+        });
 
     const { handleGestureStart, setInitialSelectedPoints, setSelectedPoint } = useGraphAtoms({
         referencePointAtom,
@@ -99,9 +92,7 @@ export const PortfolioGraph = forwardRef<PortfolioGraphRef>((_props, ref) => {
                 />
             )}
             <IgnoredNetworksBanner />
-            {showGraph && (
-                <TimeSwitch selectedTimeFrame={timeframe} onSelectTimeFrame={onSelectTimeFrame} />
-            )}
+            {showGraph && <PortfolioGraphTimeSwitch />}
         </VStack>
     );
 });
